requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
rich>=13.0.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
